+++ driver.go
@@ -25,6 +25,7 @@ import (
 	"github.com/hashicorp/nomad/plugins/shared/hclspec"
 	pstructs "github.com/hashicorp/nomad/plugins/shared/structs"
 
+	dockerref "github.com/docker/distribution/reference"
 	shelpers "github.com/hashicorp/nomad/helper/stats"
 	spec "github.com/opencontainers/runtime-spec/specs-go"
 )
@@ -474,17 +475,24 @@ func (d *Driver) StartTask(cfg *drivers.TaskConfig) (*drivers.TaskHandle, *drive
 		// FIXME: there are more variations of image sources, we should handle it
 		//        e.g. oci-archive:/... etc
 		//        see also https://github.com/hashicorp/nomad-driver-podman/issues/69
-		// do we already have this image in local storage?
-		haveImage, err := d.podman.ImageExists(d.ctx, createOpts.Image)
+
+		imageName, tag, err := parseImage(createOpts.Image)
 		if err != nil {
-			// we assume that we do NOT have a local image if something goes wrong
-			// it could work to pull/start a container
-			d.logger.Warn("failed to check for local image", "image", createOpts.Image, "err", err)
+			return nil, nil, fmt.Errorf("failed to start task, unable to parse image reference %s: %v", createOpts.Image, err)
+		}
+
+		// do we already have this image in local storage?
+		haveImage, err := d.podman.ImageExists(d.ctx, fmt.Sprintf("%s:%s", imageName, tag))
+		if err != nil {
+			d.logger.Warn("Unable to check for local image", "image", imageName, "err", err)
+			// do NOT fail this operation, instead try to pull the image
+			haveImage = false
 		}
 		if !haveImage {
+			d.logger.Debug("Pull image", "image", imageName)
 			// image is not in local storage, so we need to pull it
-			if err = d.podman.ImagePull(d.ctx, createOpts.Image); err != nil {
-				return nil, nil, fmt.Errorf("failed to start task, unable to pull image %s: %v", createOpts.Image, err)
+			if err = d.podman.ImagePull(d.ctx, imageName); err != nil {
+				return nil, nil, fmt.Errorf("failed to start task, unable to pull image %s: %v", imageName, err)
 			}
 		}
 
@@ -585,6 +593,36 @@ func memoryInBytes(strmem string) (int64, error) {
 	default:
 		return 0, fmt.Errorf("Invalid memory string: %s", strmem)
 	}
+}
+
+func parseImage(image string) (string, string, error) {
+	// strip http/https and docker transport prefix
+	for _, prefix := range []string{"http://", "https://", "docker://"} {
+		if strings.HasPrefix(image, prefix) {
+			image = strings.Replace(image, prefix, "", 1)
+		}
+	}
+
+	named, err := dockerref.ParseNormalizedNamed(image)
+	if err != nil {
+		return "", "", nil
+	}
+
+	var tag, digest string
+
+	tagged, ok := named.(dockerref.Tagged)
+	if ok {
+		tag = tagged.Tag()
+	}
+
+	digested, ok := named.(dockerref.Digested)
+	if ok {
+		digest = digested.Digest().String()
+	}
+	if len(tag) == 0 && len(digest) == 0 {
+		tag = "latest"
+	}
+	return named.Name(), tag, nil
 }
 
 // WaitTask function is expected to return a channel that will send an *ExitResult when the task